Support WebAssembly (#10)
* Factor out common parts of the WebAssembly boot logic

* Set up WebAssembly sample project

* Actually start the components on WebAssembly

* Make a separate sample page
