+++ AspNetCore.sln
@@ -1758,6 +1758,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.AspNetCore.Server
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlazorUnitedApp", "src\Components\Samples\BlazorUnitedApp\BlazorUnitedApp.csproj", "{09962BE7-48EC-44EA-B86A-3652FF0D7D1A}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlazorUnitedApp.WebAssembly", "src\Components\Samples\BlazorUnitedApp.WebAssembly\BlazorUnitedApp.WebAssembly.csproj", "{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -10551,6 +10553,22 @@ Global
 		{09962BE7-48EC-44EA-B86A-3652FF0D7D1A}.Release|x64.Build.0 = Release|Any CPU
 		{09962BE7-48EC-44EA-B86A-3652FF0D7D1A}.Release|x86.ActiveCfg = Release|Any CPU
 		{09962BE7-48EC-44EA-B86A-3652FF0D7D1A}.Release|x86.Build.0 = Release|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Debug|arm64.ActiveCfg = Debug|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Debug|arm64.Build.0 = Debug|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Debug|x64.Build.0 = Debug|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Debug|x86.Build.0 = Debug|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Release|arm64.ActiveCfg = Release|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Release|arm64.Build.0 = Release|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Release|x64.ActiveCfg = Release|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Release|x64.Build.0 = Release|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Release|x86.ActiveCfg = Release|Any CPU
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -11419,6 +11437,7 @@ Global
 		{74377D3E-E0C6-41A4-89ED-11A9C00142A9} = {166E48ED-9738-4E13-8618-0D805F6F0F65}
 		{3C7C65BF-0C13-418E-90BD-EC9C3CD282CB} = {74377D3E-E0C6-41A4-89ED-11A9C00142A9}
 		{09962BE7-48EC-44EA-B86A-3652FF0D7D1A} = {5FE1FBC1-8CE3-4355-9866-44FE1307C5F1}
+		{42DA21DE-8DC8-4BED-BCE8-D7DDA9CEC60D} = {5FE1FBC1-8CE3-4355-9866-44FE1307C5F1}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {3E8720B3-DBDD-498C-B383-2CC32A054E8F}
