+++ src/Components/Samples/BlazorUnitedApp.WebAssembly/_Imports.razor
@@ -1,0 +1,7 @@
+﻿@using System.Net.Http
+@using System.Net.Http.Json
+@using Microsoft.AspNetCore.Components.Routing
+@using Microsoft.AspNetCore.Components.Web
+@using Microsoft.AspNetCore.Components.WebAssembly.Http
+@using Microsoft.JSInterop
+@using BlazorUnitedApp.WebAssembly
