+++ src/Components/Samples/BlazorUnitedApp/Pages/CounterServer.razor
@@ -2,6 +2,6 @@
 
 <PageTitle>Counter</PageTitle>
 
-<h1>Counter</h1>
+<h1>Counter (interactive on server)</h1>
 
 <CounterUI InitialCount="100" rendermode="@WebComponentRenderMode.Server" />
