+++ src/Components/Samples/BlazorUnitedApp/Pages/CounterWebAssembly.razor
@@ -1,0 +1,7 @@
+﻿@page "/counter-webassembly"
+
+<PageTitle>Counter</PageTitle>
+
+<h1>Counter (interactive on WebAssembly)</h1>
+
+<CounterUI InitialCount="100" rendermode="@WebComponentRenderMode.WebAssembly" />
