+++ src/Components/Samples/BlazorUnitedApp/Shared/NavMenu.razor
@@ -23,6 +23,11 @@
             </NavLink>
         </li>
         <li class="nav-item px-3">
+            <NavLink class="nav-link" href="counter-webassembly">
+                <span class="oi oi-plus" aria-hidden="true"></span> Counter (wasm)
+            </NavLink>
+        </li>
+        <li class="nav-item px-3">
             <NavLink class="nav-link" href="fetchdata">
                 <span class="oi oi-list-rich" aria-hidden="true"></span> Fetch data
             </NavLink>
