+++ src/Components/Samples/BlazorUnitedApp/_Imports.razor
@@ -7,3 +7,4 @@
 @using Microsoft.JSInterop
 @using BlazorUnitedApp
 @using BlazorUnitedApp.Shared
+@using BlazorUnitedApp.WebAssembly
