+++ src/Components/Web.JS/src/Boot.United.ts
@@ -5,7 +5,8 @@ import { Blazor } from './GlobalExports';
 import { shouldAutoStart } from './BootCommon';
 import { UnitedStartOptions } from './Platform/Circuits/UnitedStartOptions';
 import { startCircuit } from './Boot.Server.Common';
-import { discoverComponents, ServerComponentDescriptor } from './Services/ComponentDescriptorDiscovery';
+import { discoverComponents, ServerComponentDescriptor, WebAssemblyComponentDescriptor } from './Services/ComponentDescriptorDiscovery';
+import { startWebAssembly } from './Boot.WebAssembly.Common';
 
 let started = false;
 
@@ -15,12 +16,22 @@ async function boot(options?: Partial<UnitedStartOptions>): Promise<void> {
   }
   started = true;
 
+  const serverComponents = discoverComponents(document, 'server') as ServerComponentDescriptor[];
+  const webAssemblyComponents = discoverComponents(document, 'webassembly') as WebAssemblyComponentDescriptor[];
+
+  if (serverComponents.length && webAssemblyComponents.length) {
+    throw new Error('TODO: Support having both Server and WebAssembly components at the same time. Not doing that currently as it overlaps with a different prototype.');
+  }
+
   // Only set up a circuit if the page actually contains interactive server components
   // Later on, when we add progressive enhancement to navigation, we will also want to
   // auto-close circuits when the last root component is removed.
-  const components = discoverComponents(document, 'server') as ServerComponentDescriptor[];
-  if (components.length) {
+  if (serverComponents.length) {
     await startCircuit(options?.circuit);
+  }
+
+  if (webAssemblyComponents.length) {
+    await startWebAssembly(options?.webAssembly);
   }
 }
